pyproject.toml
[tool.poetry]
name        = "mgraph_ai_service_graph"
version     = "v1.1.10"
description = "MGraph-AI__Service__Graph"
authors     = ["Dinis Cruz <dinis.cruz@owasp.org>"]
license     = "Apache 2.0"
readme      = "README.md"
homepage    = "https://github.com/the-cyber-boardroom/MGraph-AI__Service__Graph"
repository  = "https://github.com/the-cyber-boardroom/MGraph-AI__Service__Graph"

[tool.poetry.dependencies]
python                         = "^3.12"
osbot-fast-api-serverless      = "*"
mgraph-db                      = "*"
mgraph-ai-service-cache-client = "*"
httpx                          = "^0.28.1"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"